% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadChinaMap}
\alias{loadChinaMap}
\title{load China Map data only once.}
\usage{
loadChinaMap()
}
\value{
SpatialPolygonsDataFrame china map
}
\description{
load China Map data only once.
}
\examples{
\dontrun{
   loadChinaMap()
}
}
\keyword{map}

